requirements.txt
Django==2.2.10
djangorestframework==3.10.2
djangorestframework-stubs==0.4.2
pytz==2018.9
python-dateutil==2.7.3
six==1.11.0
requests==2.20.0
setuptools==51.1.2
passlib==1.7.1
urllib3==1.24.2
django-bootstrap4==1.1.1
asn1crypto==1.3.0
cffi==1.14.0
pycparser==2.20
cryptography==3.3.1